pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "poker-agent"
version = "0.1.0"
description = "Autonomous poker-playing agent for Telegram mini-app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "mss>=9.0.0",
    "pyautogui>=0.9.54",
    "Pillow>=10.0.0",
    "treys>=0.1.8",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
    "transitions>=0.9.0",
    "typer>=0.12.0",
    "imagehash>=4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
poker-agent = "poker_agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/poker_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
